pyproject.toml
[project]
name = "notes-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "line-bot-sdk>=3.21.0",
    "notion-client>=2.7.0",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "google-auth>=2.29.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.119.0",
]
